@@ -3,17 +3,22 @@ version: 0.6.0
 description: Revision calculator for git projects
 
 environment:
-  sdk: '>=2.0.0-dev.61.0 <3.0.0'
+  sdk: '>=2.15.0 <3.0.0'
 
 dependencies:
-  args: '^1.2.0'
+  args: ^2.3.0
 
 dev_dependencies:
-  test: '^1.2.0'
-  mockito: '^3.0.0-beta+1'
-  dart_style: '^1.1.2'
-  path: ^1.6.2
-  archive: '>=1.0.0 <3.0.0'
+  archive: ^3.0.0
+  collection: ^1.15.0
+  dart_style: ^2.2.0
+  http: ^0.13.4
+  lint: ^1.0.0
+  mockito: ^5.0.0
+  path: ^1.8.1
+  pub_semver: ^2.1.0
+  test: ^1.20.0
+  yaml: ^3.1.0
 
 executables:
   # a valid dart filename doesn't contain a dash -> remapping
