@@ -8,17 +8,22 @@ final DateTime initTime = DateTime.utc(2017, DateTime.january, 10);
 
 void main() {
   group('initialize', () {
-    TempGit git;
+    late TempGit git;
     setUp(() async {
       git = await makeTempGit();
     });
 
     test('no git', () async {
-      await git.run(name: 'init commit', script: sh("""
+      await git.run(
+        name: 'init commit',
+        script: sh(
+          """
           echo 'Hello World' > a.txt
-          """));
+          """,
+        ),
+      );
 
-      var out = await git.revision(['--full']);
+      final out = await git.revision(['--full']);
 
       expect(out, contains('versionCode: 0\n'));
       expect(out, contains('versionName: 0_0000000-dirty\n'));
@@ -38,12 +43,17 @@ void main() {
     });
 
     test('no commmit', () async {
-      await git.run(name: 'init commit', script: sh("""
+      await git.run(
+        name: 'init commit',
+        script: sh(
+          """
           git init
           echo 'Hello World' > a.txt
-          """));
+          """,
+        ),
+      );
 
-      var out = await git.revision(['--full']);
+      final out = await git.revision(['--full']);
 
       expect(out, contains('versionCode: 0\n'));
       expect(out, contains('versionName: 0_0000000-dirty\n'));
@@ -64,7 +74,7 @@ void main() {
   });
 
   group('orphan branch', () {
-    TempGit git;
+    late TempGit git;
     setUp(() async {
       git = await makeTempGit();
     });
@@ -72,7 +82,10 @@ void main() {
     test('first commit', () async {
       git.skipCleanup = true;
       print('cd ${git.repo.path} && fork');
-      await git.run(name: 'init with 3 commits', script: sh("""
+      await git.run(
+        name: 'init with 3 commits',
+        script: sh(
+          """
           git init
           echo 'Hello World' > a.txt
           git add a.txt
@@ -84,22 +97,30 @@ void main() {
           
           echo 'third commit' > a.txt
           ${commit("third commit", initTime.add(day))}
-          """));
+          """,
+        ),
+      );
 
-      await git.run(name: 'create orphan commit', script: sh("""
+      await git.run(
+        name: 'create orphan commit',
+        script: sh(
+          """
           git checkout --orphan another_root
           echo 'World Hello' > a.txt
           git add a.txt
           
           ${commit("orphan commit", initTime.add(day * 2))}
-          """));
+          """,
+        ),
+      );
 
-      var out = await git.revision(['--full']);
+      final out = await git.revision(['--full']);
 
       expect(out, contains('versionCode: 0\n'));
-      expect(out, contains('versionName: 0_another_root+1_2c8318a\n'));
+      expect(out, contains('versionName: 0_another_root+1_0084417\n'));
       expect(out, contains('baseBranch: master\n'));
       expect(out, contains('currentBranch: another_root\n'));
+      expect(out, contains('sha1: 008441704a1f59649f6dbce4487f9a8747edf124\n'));
       expect(out, contains('completeFirstOnlyBaseBranchCommitCount: 3\n'));
       expect(out, contains('baseBranchCommitCount: 0\n'));
       expect(out, contains('baseBranchTimeComponent: 0\n'));
@@ -113,7 +134,10 @@ void main() {
     test('3 commits', () async {
       git.skipCleanup = true;
       print('cd ${git.repo.path} && fork');
-      await git.run(name: 'init with 3 commits', script: sh("""
+      await git.run(
+        name: 'init with 3 commits',
+        script: sh(
+          """
           git init
           echo 'Hello World' > a.txt
           git add a.txt
@@ -125,9 +149,14 @@ void main() {
           
           echo 'third commit' > a.txt
           ${commit("third commit", initTime.add(day))}
-          """));
+          """,
+        ),
+      );
 
-      await git.run(name: 'create orphan commit', script: sh("""
+      await git.run(
+        name: 'create orphan commit',
+        script: sh(
+          """
           git checkout --orphan another_root
           echo 'World Hello' > a.txt
           git add a.txt
@@ -139,12 +168,14 @@ void main() {
           
           echo 'third commit' > a.txt
           ${commit("third commit", initTime.add(day * 3))}
-          """));
+          """,
+        ),
+      );
 
-      var out = await git.revision(['--full']);
+      final out = await git.revision(['--full']);
 
       expect(out, contains('versionCode: 0\n'));
-      expect(out, contains('versionName: 0_another_root+3_c3181a8\n'));
+      expect(out, contains('versionName: 0_another_root+3_f2d44aa\n'));
       expect(out, contains('baseBranch: master\n'));
       expect(out, contains('currentBranch: another_root\n'));
       expect(out, contains('completeFirstOnlyBaseBranchCommitCount: 3\n'));
@@ -159,24 +190,29 @@ void main() {
   });
 
   group('master only', () {
-    TempGit git;
+    late TempGit git;
     setUp(() async {
       git = await makeTempGit();
     });
 
     test('first commmit', () async {
-      await git.run(name: 'init commit', script: sh("""
+      await git.run(
+        name: 'init commit',
+        script: sh(
+          """
           git init
           echo 'Hello World' > a.txt
           git add a.txt
           
           ${commit("initial commit", initTime)}
-          """));
+          """,
+        ),
+      );
 
-      var out = await git.revision(['--full']);
+      final out = await git.revision(['--full']);
 
       expect(out, contains('versionCode: 1\n'));
-      expect(out, contains('versionName: 1_4814a87\n'));
+      expect(out, contains('versionName: 1_b5f9bcd\n'));
       expect(out, contains('baseBranch: master\n'));
       expect(out, contains('currentBranch: master\n'));
       expect(out, contains('completeFirstOnlyBaseBranchCommitCount: 1\n'));
@@ -189,7 +225,10 @@ void main() {
     });
 
     test('3 commits', () async {
-      await git.run(name: 'init with 3 commits', script: sh("""
+      await git.run(
+        name: 'init with 3 commits',
+        script: sh(
+          """
           git init
           echo 'Hello World' > a.txt
           git add a.txt
@@ -201,12 +240,14 @@ void main() {
           
           echo 'third commit' > a.txt
           ${commit("third commit", initTime.add(day))}
-          """));
+          """,
+        ),
+      );
 
-      var out = await git.revision(['--full']);
+      final out = await git.revision(['--full']);
 
       expect(out, contains('versionCode: 6\n'));
-      expect(out, contains('versionName: 6_937de8c\n'));
+      expect(out, contains('versionName: 6_f3aacda\n'));
       expect(out, contains('baseBranch: master\n'));
       expect(out, contains('currentBranch: master\n'));
       expect(out, contains('completeFirstOnlyBaseBranchCommitCount: 3\n'));
@@ -218,7 +259,10 @@ void main() {
     });
 
     test("merge branch with old commits doesn't increase the revision of previous commits", () async {
-      await git.run(name: 'init master branch', script: sh("""
+      await git.run(
+        name: 'init master branch',
+        script: sh(
+          """
           git init
           echo 'Hello World' > a.txt
           git add a.txt
@@ -226,13 +270,18 @@ void main() {
           
           echo 'second commit' > a.txt
           ${commit("second commit", initTime.add(hour * 6))}
-          """));
+          """,
+        ),
+      );
 
       // get current revision, should not change afterwards
-      var out1 = await git.revision(['--full']);
+      final out1 = await git.revision(['--full']);
       expect(out1, contains('versionCode: 3'));
 
-      await git.run(name: 'merge feature B', script: sh("""
+      await git.run(
+        name: 'merge feature B',
+        script: sh(
+          """
           # branch from initial commit
           git checkout HEAD^1
           git checkout -b 'featureB'
@@ -247,35 +296,45 @@ void main() {
           
           git checkout master
           ${merge("featureB", initTime.add(day * 2 + (hour * 1)))}
-          """));
+          """,
+        ),
+      );
 
       // revision obviously increased after merge
-      var out2 = await git.revision(['--full']);
+      final out2 = await git.revision(['--full']);
       expect(out2, contains('baseBranchTimeComponent: 6\n'));
       expect(out2, contains('baseBranchCommitCount: 5\n'));
       expect(out2, contains('versionCode: 11\n'));
-      expect(out2, contains('versionName: 11_7810fd1\n'));
+      expect(out2, contains('versionName: 11_99c52ee\n'));
 
-      await git.run(name: 'go back to commit before merge', script: sh("""
+      await git.run(
+        name: 'go back to commit before merge',
+        script: sh(
+          """
           git checkout master
           git checkout HEAD^1
-          """));
+          """,
+        ),
+      );
 
       // same revision as before
-      var out3 = await git.revision(['--full']);
+      final out3 = await git.revision(['--full']);
       expect(out3, contains('versionCode: 3\n'));
-      expect(out3, contains('versionName: 3_046e78b\n'));
+      expect(out3, contains('versionName: 3_51f6726\n'));
     });
   });
 
   group('feature branch', () {
-    TempGit git;
+    late TempGit git;
     setUp(() async {
       git = await makeTempGit();
     });
 
     test("no branch name - fallback to sha1", () async {
-      await git.run(name: 'init master branch - work on featureB', script: sh("""
+      await git.run(
+        name: 'init master branch - work on featureB',
+        script: sh(
+          """
           git init
           echo 'Hello World' > a.txt
           git add a.txt
@@ -293,20 +352,30 @@ void main() {
           
           echo 'fix bug' > b.txt
           ${commit("fix bug", initTime.add(day))}
-          """));
+          """,
+        ),
+      );
 
-      await git.run(name: 'delete feature branch and stay on detached head', script: sh("""
+      await git.run(
+        name: 'delete feature branch and stay on detached head',
+        script: sh(
+          """
           git checkout master
           git branch -D featureB
-          git checkout f73f20d
-          """));
+          git checkout 331b280
+          """,
+        ),
+      );
 
-      var out = await git.revision(['--full']);
-      expect(out, contains('versionName: 3+2_f73f20d\n'));
+      final out = await git.revision(['--full']);
+      expect(out, contains('versionName: 3+2_331b280\n'));
     });
 
     test("feature branch still has +commits after merge in master", () async {
-      await git.run(name: 'init master branch - work on featureB', script: sh("""
+      await git.run(
+        name: 'init master branch - work on featureB',
+        script: sh(
+          """
           git init
           echo 'Hello World' > a.txt
           git add a.txt
@@ -324,12 +393,17 @@ void main() {
           
           echo 'fix bug' > b.txt
           ${commit("fix bug", initTime.add(day))}
-          """));
+          """,
+        ),
+      );
 
-      var out = await git.revision(['--full']);
-      expect(out, contains('versionName: 3_featureB+2_f73f20d\n'));
+      final out = await git.revision(['--full']);
+      expect(out, contains('versionName: 3_featureB+2_331b280\n'));
 
-      await git.run(name: 'continue work on master and merge featureB', script: sh("""
+      await git.run(
+        name: 'continue work on master and merge featureB',
+        script: sh(
+          """
           git checkout master
           echo 'third commit' > a.txt
           ${commit("third commit", initTime.add(day + (hour * 2)))}
@@ -339,15 +413,20 @@ void main() {
           
           # Go back to feature branch
           git checkout featureB
-      """));
+      """,
+        ),
+      );
 
       // back on featureB the previous output should not change
-      var out2 = await git.revision(['--full']);
-      expect(out2, contains('versionName: 3_featureB+2_f73f20d\n'));
+      final out2 = await git.revision(['--full']);
+      expect(out2, contains('versionName: 3_featureB+2_331b280\n'));
     });
 
     test("git flow - baseBranch=develop - merge develop -> master increases revision", () async {
-      await git.run(name: 'init master branch - create develop', script: sh("""
+      await git.run(
+        name: 'init master branch - create develop',
+        script: sh(
+          """
           git init
           echo 'Hello World' > a.txt
           git add a.txt
@@ -358,9 +437,14 @@ void main() {
           
           # Create develop branch
           git checkout -b 'develop'
-          """));
+          """,
+        ),
+      );
 
-      await git.run(name: 'work on feature B', script: sh("""
+      await git.run(
+        name: 'work on feature B',
+        script: sh(
+          """
           git checkout develop
           git checkout -b 'featureB'
           echo 'implement feature B' > b.txt
@@ -370,9 +454,14 @@ void main() {
           
           echo 'fix bug' > b.txt
           ${commit("fix bug", initTime.add(day))}
-          """));
+          """,
+        ),
+      );
 
-      await git.run(name: 'work on feature C', script: sh("""
+      await git.run(
+        name: 'work on feature C',
+        script: sh(
+          """
           git checkout develop
           git checkout -b 'featureC'
           echo 'implement feature C' > c.txt
@@ -381,9 +470,14 @@ void main() {
           
           echo 'fix a bug' > c.txt
           ${commit("fix bug", initTime.add(day * 2))}
-          """));
+          """,
+        ),
+      );
 
-      await git.run(name: 'work on feature D', script: sh("""
+      await git.run(
+        name: 'work on feature D',
+        script: sh(
+          """
           git checkout develop
           git checkout -b 'featureD'
           echo 'implement feature D' > d.txt
@@ -392,41 +486,57 @@ void main() {
           
           echo 'fix more bugs' > d.txt
           ${commit("fix bug", initTime.add(day + (hour * 5)))}
-          """));
+          """,
+        ),
+      );
 
-      await git.run(name: 'merge C then B into develop and release to master', script: sh("""
+      await git.run(
+        name: 'merge C then B into develop and release to master',
+        script: sh(
+          """
           git checkout develop
           ${merge('featureC', initTime.add(day * 2 + (hour * 1)))}
           ${merge('featureB', initTime.add(day * 2 + (hour * 2)))}
           
           git checkout master
           ${merge("develop", initTime.add(day * 2 + (hour * 3)))}
-          """));
+          """,
+        ),
+      );
 
-      await git.run(name: 'merge D into develop and release to mastser', script: sh("""
+      await git.run(
+        name: 'merge D into develop and release to mastser',
+        script: sh(
+          """
           git checkout develop
           ${merge("featureD", initTime.add(day * 2 + (hour * 4)))}
           
           git checkout master
           ${merge("develop", initTime.add(day * 2 + (hour * 5)))}
-          """));
+          """,
+        ),
+      );
 
       // master should be only +2 ahead which are the two merge commits (develop -> master)
       // master will always +1 ahead of develop even when merging (master -> develop)
-      var out2 = await git.revision(['--full', '--baseBranch', 'develop']);
-      expect(out2, contains('versionName: 17_master+2_127e54e\n'));
+      final out2 = await git.revision(['--full', '--baseBranch', 'develop']);
+      expect(out2, contains('versionName: 17_master+2_a658ff8\n'));
     });
   });
 
   group('remote', () {
-    TempGit git;
+    late TempGit git;
     setUp(() async {
       git = await makeTempGit();
     });
 
     test("master only on remote", () async {
-      var repo2 = await io.Directory("${git.root.path}${io.Platform.pathSeparator}remoteRepo").create();
-      await git.run(name: 'init master branch', repo: repo2, script: sh("""
+      final repo2 = await io.Directory("${git.root.path}${io.Platform.pathSeparator}remoteRepo").create();
+      await git.run(
+        name: 'init master branch',
+        repo: repo2,
+        script: sh(
+          """
           git init
           echo 'Hello World' > a.txt
           git add a.txt
@@ -434,9 +544,14 @@ void main() {
           
           echo 'second commit' > a.txt
           ${commit("second commit", initTime.add(hour * 4))}
-          """));
+          """,
+        ),
+      );
 
-      await git.run(name: 'clone and implement feature B', script: sh("""
+      await git.run(
+        name: 'clone and implement feature B',
+        script: sh(
+          """
           git clone ${repo2.path} .
           git checkout -b 'featureB'
           echo 'implement feature B' > b.txt
@@ -446,22 +561,28 @@ void main() {
           
           echo 'fix bug' > b.txt
           ${commit("fix bug", initTime.add(day))}
-          """));
+          """,
+        ),
+      );
 
-      var out = await git.revision(['--full']);
-      expect(out, contains('versionName: 2_featureB+2_3067399\n'));
+      final out = await git.revision(['--full']);
+      expect(out, contains('versionName: 2_featureB+2_d3e1844\n'));
 
       // now master branch is only available on remote
       await git.run(name: 'delete master branch', script: "git branch -d master");
 
       // output is unchanged
-      var out2 = await git.revision(['--full']);
-      expect(out2, contains('versionName: 2_featureB+2_3067399\n'));
+      final out2 = await git.revision(['--full']);
+      expect(out2, contains('versionName: 2_featureB+2_d3e1844\n'));
     });
 
     test("master only on remote which is not called origin", () async {
-      var repo2 = await io.Directory("${git.root.path}${io.Platform.pathSeparator}remoteRepo").create();
-      await git.run(name: 'init master branch', repo: repo2, script: sh("""
+      final repo2 = await io.Directory("${git.root.path}${io.Platform.pathSeparator}remoteRepo").create();
+      await git.run(
+        name: 'init master branch',
+        repo: repo2,
+        script: sh(
+          """
           git init
           echo 'Hello World' > a.txt
           git add a.txt
@@ -469,9 +590,14 @@ void main() {
           
           echo 'second commit' > a.txt
           ${commit("second commit", initTime.add(hour * 4))}
-          """));
+          """,
+        ),
+      );
 
-      await git.run(name: 'clone and implement feature B', script: sh("""
+      await git.run(
+        name: 'clone and implement feature B',
+        script: sh(
+          """
           git clone -o first ${repo2.path} .
           git checkout -b 'featureB'
           echo 'implement feature B' > b.txt
@@ -480,22 +606,28 @@ void main() {
           
           echo 'fix bug' > b.txt
           ${commit("fix bug", initTime.add(day))}
-          """));
+          """,
+        ),
+      );
 
-      var out = await git.revision(['--full']);
-      expect(out, contains('versionName: 2_featureB+2_3067399\n'));
+      final out = await git.revision(['--full']);
+      expect(out, contains('versionName: 2_featureB+2_d3e1844\n'));
 
       // now master branch is only available on remote
       await git.run(name: 'delete master branch', script: "git branch -d master");
 
       // output is unchanged
-      var out2 = await git.revision(['--full']);
-      expect(out2, contains('versionName: 2_featureB+2_3067399\n'));
+      final out2 = await git.revision(['--full']);
+      expect(out2, contains('versionName: 2_featureB+2_d3e1844\n'));
     });
 
     test("master only on one remote - multiple remotes", () async {
-      var repo2 = await io.Directory("${git.root.path}${io.Platform.pathSeparator}remoteRepo").create();
-      await git.run(name: 'init master branch', repo: repo2, script: sh("""
+      final repo2 = await io.Directory("${git.root.path}${io.Platform.pathSeparator}remoteRepo").create();
+      await git.run(
+        name: 'init master branch',
+        repo: repo2,
+        script: sh(
+          """
           git init
           echo 'Hello World' > a.txt
           git add a.txt
@@ -503,9 +635,14 @@ void main() {
           
           echo 'second commit' > a.txt
           ${commit("second commit", initTime.add(hour * 4))}
-          """));
+          """,
+        ),
+      );
 
-      await git.run(name: 'add remotes and start working', script: sh("""
+      await git.run(
+        name: 'add remotes and start working',
+        script: sh(
+          """
           git init
           git remote add first ${repo2.path}
           git remote add second ${repo2.path}
@@ -518,17 +655,19 @@ void main() {
           echo 'implement feature B' > b.txt
           git add b.txt
           ${commit("implement feature B", initTime.add(hour * 6))}
-          """));
+          """,
+        ),
+      );
 
-      var out = await git.revision(['--full']);
-      expect(out, contains('versionName: 2_featureB+1_e45ede3'));
+      final out = await git.revision(['--full']);
+      expect(out, contains('versionName: 2_featureB+1_9006da1'));
 
       // now master branch is only available on remote
       await git.run(name: 'delete master branch', script: "git branch -d master");
 
       // output is unchanged
-      var out2 = await git.revision(['--full']);
-      expect(out2, contains('versionName: 2_featureB+1_e45ede3\n'));
+      final out2 = await git.revision(['--full']);
+      expect(out2, contains('versionName: 2_featureB+1_9006da1\n'));
     });
   });
 }
